requirements.txt
anyio==4.6.0                                                                                                                          
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
Django==5.1.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
firebase-admin==6.5.0
gcloud==0.18.3
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
jwcrypto==1.5.6
MarkupSafe==2.1.5
msgpack==1.1.0
nltk==3.9.1
numpy==2.1.1
oauth2client==4.1.3
openapi-codec==1.3.2
packaging==24.1
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyparsing==3.1.4
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20